Workaround scavetool bug

+++ simulations/simulate.sh
@@ -10,7 +10,10 @@ VECTORS=true
 INIFILE=simulations.ini
 REPEAT=
 OUTNAME=data
-while getopts s:c:o:r:tn flag
+COUNT=1000
+EXPORTONLY=false
+
+while getopts s:c:o:r:tnx flag
 do
 	case "${flag}" in
 		s) SCENARIO=${OPTARG};;
@@ -19,6 +22,8 @@ do
 		r) REPEAT=--repeat=${OPTARG};;
 		t) INIFILE=tests.ini;;
 		n) VECTORS=false;;
+		b) COUNT=${OPTARG};;
+		x) EXPORTONLY=true;;
 	esac
 done
 
@@ -27,15 +32,43 @@ if [ -z "$SCENARIO" ] || [ -z "$CONFIGNAME" ]; then
 	exit
 fi
 
-rm -rf $RESULTDIR/$CONFIGNAME
-
-$RUNNER -j $(nproc) ../src/pecsn -n .:../src -c $CONFIGNAME -f $INIFILE -u Cmdenv $REPEAT
+if [ "$EXPORTONLY" = false ]; then
+	rm -rf $RESULTDIR/$CONFIGNAME
+	$RUNNER -j $(nproc) ../src/pecsn -n .:../src -c $CONFIGNAME -f $INIFILE -u Cmdenv $REPEAT
+fi
 
 mkdir -p $ANALYSISDIR/$SCENARIO/$EXPDATAFOLDER
 if [ "$VECTORS" = true ]; then
-	$SCAVETOOL x $RESULTDIR/$CONFIGNAME/*.{sca,vec} -o $ANALYSISDIR/$SCENARIO/$EXPDATAFOLDER/$OUTNAME.csv
+	INFILES=($RESULTDIR/$CONFIGNAME/*.{sca,vec})
 else
-	$SCAVETOOL x $RESULTDIR/$CONFIGNAME/*.sca -o $ANALYSISDIR/$SCENARIO/$EXPDATAFOLDER/$OUTNAME.csv
+	INFILES=($RESULTDIR/$CONFIGNAME/*.sca)
 fi
+
+IFS=$'\n' INFILES=($(sort <<<"${INFILES[*]}"))
+unset IFS
+
+rm -f $ANALYSISDIR/$SCENARIO/$EXPDATAFOLDER/$OUTNAME.csv
+
+# Extract data from result files in batches
+# Needed due to a bug in scavetool that is unable to process large number of
+# files :(
+echo -n 'Exporting data...'
+OPFILES=()
+i=0
+while ((${#INFILES[@]})); do
+	CURFILES=("${INFILES[@]:0:$COUNT}")
+	$SCAVETOOL x -F CSV-R -x columnNames=false ${CURFILES[@]} -o - > /tmp/${SCENARIO}_${OUTNAME}$i.csv
+	OPFILES+=("/tmp/${SCENARIO}_${OUTNAME}$i.csv")
+	i+=1
+	INFILES=("${INFILES[@]:$COUNT}")
+done
+echo 'done!'
+if [ "$VECTORS" = true ]; then
+	echo 'run,type,module,name,attrname,attrvalue,value,count,sumweights,mean,stddev,min,max,binedges,binvalues,vectime,vecvalue' > $ANALYSISDIR/$SCENARIO/$EXPDATAFOLDER/$OUTNAME.csv
+else
+	echo 'run,type,module,name,attrname,attrvalue,value,count,sumweights,mean,stddev,min,max,binedges,binvalues' > $ANALYSISDIR/$SCENARIO/$EXPDATAFOLDER/$OUTNAME.csv
+fi
+cat ${OPFILES[@]} >> $ANALYSISDIR/$SCENARIO/$EXPDATAFOLDER/$OUTNAME.csv
+rm "${OPFILES[@]}"
 
 echo "DONE! Output written to $ANALYSISDIR/$SCENARIO/$EXPDATAFOLDER/$OUTNAME.csv"
